+++ Example/AEXMLDemo/ViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import AEXML
 
 class ViewController: UIViewController {
 
@@ -20,7 +21,10 @@ class ViewController: UIViewController {
         guard let
             xmlPath = NSBundle.mainBundle().pathForResource("example", ofType: "xml"),
             data = NSData(contentsOfFile: xmlPath)
-        else { return }
+        else {
+            print("resource not found!")
+            return
+        }
         
         // example of using NSXMLParserOptions
         var options = AEXMLDocument.NSXMLParserOptions()
@@ -89,8 +93,11 @@ class ViewController: UIViewController {
             // prints Siberian
             print(xmlDoc.root["cats"]["cat"].attributes["breed"]!)
             
-            // prints element <badexample> not found
-            print(xmlDoc["badexample"]["notexisting"].stringValue)
+            // prints <cat breed="Siberian" color="lightgray">Tinna</cat>
+            print(xmlDoc.root["cats"]["cat"].xmlStringCompact)
+            
+            // prints Optional(AEXMLExample.AEXMLElement.Error.ElementNotFound)
+            print(xmlDoc["NotExistingElement"].error)
         }
         catch {
             print("\(error)")