+++ Sources/AEXML.swift
@@ -24,13 +24,26 @@
 
 import Foundation
 
+// MARK: - AEXMLElement
+
 /**
     This is base class for holding XML structure.
 
     You can access its structure by using subscript like this: `element["foo"]["bar"]` which would
     return `<bar></bar>` element from `<element><foo><bar></bar></foo></element>` XML as an `AEXMLElement` object.
 */
-public class AEXMLElement: NSObject {
+public class AEXMLElement {
+    
+    /// A type representing an error value that can be inside `error` property.
+    public enum Error: ErrorType {
+        case ElementNotFound
+        case RootElementMissing
+    }
+    
+    private struct Defaults {
+        static let name = String()
+        static let attributes = [String : String]()
+    }
     
     // MARK: Properties
     
@@ -49,22 +62,11 @@ public class AEXMLElement: NSObject {
     /// XML Element attributes (defaults to empty dictionary).
     public var attributes: [String : String]
     
+    /// Error value (`nil` if there is no error).
+    public var error: Error?
+    
     /// String representation of `value` property (if `value` is `nil` this is empty String).
     public var stringValue: String { return value ?? String() }
-    
-    /// String representation of `value` property with special characters escaped (if `value` is `nil` this is empty String).
-    public var escapedStringValue: String {
-        // we need to make sure "&" is escaped first. Not doing this may break escaping the other characters
-        var escapedString = stringValue.stringByReplacingOccurrencesOfString("&", withString: "&amp;", options: .LiteralSearch)
-        
-        // replace the other four special characters
-        let escapeChars = ["<" : "&lt;", ">" : "&gt;", "'" : "&apos;", "\"" : "&quot;"]
-        for (char, echar) in escapeChars {
-            escapedString = escapedString.stringByReplacingOccurrencesOfString(char, withString: echar, options: .LiteralSearch)
-        }
-        
-        return escapedString
-    }
     
     /// Boolean representation of `value` property (if `value` is "true" or 1 this is `True`, otherwise `False`).
     public var boolValue: Bool { return stringValue.lowercaseString == "true" || Int(stringValue) == 1 ? true : false }
@@ -75,21 +77,16 @@ public class AEXMLElement: NSObject {
     /// Double representation of `value` property (this is **0.00** if `value` can't be represented as Double).
     public var doubleValue: Double { return (stringValue as NSString).doubleValue }
     
-    private struct Defaults {
-        static let name = String()
-        static let attributes = [String : String]()
-    }
-    
     // MARK: Lifecycle
     
     /**
         Designated initializer - all parameters are optional.
     
-        :param: name XML element name.
-        :param: value XML element value
-        :param: attributes XML element attributes
+        - parameter name: XML element name.
+        - parameter value: XML element value
+        - parameter attributes: XML element attributes
     
-        :returns: An initialized `AEXMLElement` object.
+        - returns: An initialized `AEXMLElement` object.
     */
     public init(_ name: String? = nil, value: String? = nil, attributes: [String : String]? = nil) {
         self.name = name ?? Defaults.name
@@ -99,18 +96,16 @@ public class AEXMLElement: NSObject {
     
     // MARK: XML Read
     
-    /// This element name is used when unable to find element.
-    public static let errorElementName = "AEXMLError"
-    
-    // The first element with given name **(AEXMLError element if not exists)**.
+    /// The first element with given name **(Empty element with error if not exists)**.
     public subscript(key: String) -> AEXMLElement {
-        if name == AEXMLElement.errorElementName {
-            return self
-        } else {
-            let filtered = children.filter { $0.name == key }
-            let errorElement = AEXMLElement(AEXMLElement.errorElementName, value: "element <\(key)> not found")
-            return filtered.count > 0 ? filtered.first! : errorElement
+        guard let
+            first = children.filter({ $0.name == key }).first
+        else {
+            let errorElement = AEXMLElement(key)
+            errorElement.error = Error.ElementNotFound
+            return errorElement
         }
+        return first
     }
     
     /// Returns all of the elements with equal name as `self` **(nil if not exists)**.
@@ -142,9 +137,9 @@ public class AEXMLElement: NSObject {
     /**
         Returns all elements with given value.
         
-        :param: value XML element value.
+        - parameter value: XML element value.
         
-        :returns: Optional Array of found XML elements.
+        - returns: Optional Array of found XML elements.
     */
     public func allWithValue(value: String) -> [AEXMLElement]? {
         let found = allWithCondition { (element) -> Bool in
@@ -156,9 +151,9 @@ public class AEXMLElement: NSObject {
     /**
         Returns all elements with given attributes.
     
-        :param: attributes Dictionary of Keys and Values of attributes.
+        - parameter attributes: Dictionary of Keys and Values of attributes.
     
-        :returns: Optional Array of found XML elements.
+        - returns: Optional Array of found XML elements.
     */
     public func allWithAttributes(attributes: [String : String]) -> [AEXMLElement]? {
         let found = allWithCondition { (element) -> Bool in
@@ -178,9 +173,9 @@ public class AEXMLElement: NSObject {
     /**
         Adds child XML element to `self`.
     
-        :param: child Child XML element to add.
+        - parameter child: Child XML element to add.
     
-        :returns: Child XML element with `self` as `parent`.
+        - returns: Child XML element with `self` as `parent`.
     */
     public func addChild(child: AEXMLElement) -> AEXMLElement {
         child.parent = self
@@ -191,11 +186,11 @@ public class AEXMLElement: NSObject {
     /**
         Adds child XML element to `self`.
         
-        :param: name Child XML element name.
-        :param: value Child XML element value.
-        :param: attributes Child XML element attributes.
+        - parameter name: Child XML element name.
+        - parameter value: Child XML element value.
+        - parameter attributes: Child XML element attributes.
         
-        :returns: Child XML element with `self` as `parent`.
+        - returns: Child XML element with `self` as `parent`.
     */
     public func addChild(name name: String, value: String? = nil, attributes: [String : String]? = nil) -> AEXMLElement {
         let child = AEXMLElement(name, value: value, attributes: attributes)
@@ -208,7 +203,7 @@ public class AEXMLElement: NSObject {
     }
     
     private func removeChild(child: AEXMLElement) {
-        if let childIndex = children.indexOf(child) {
+        if let childIndex = children.indexOf({ $0 === child }) {
             children.removeAtIndex(childIndex)
         }
     }
@@ -223,12 +218,15 @@ public class AEXMLElement: NSObject {
         return count
     }
     
-    private func indentation(var count: Int) -> String {
+    private func indentation(depth: Int) -> String {
+        var count = depth
         var indent = String()
+        
         while count > 0 {
             indent += "\t"
             count -= 1
         }
+        
         return indent
     }
     
@@ -243,7 +241,7 @@ public class AEXMLElement: NSObject {
         if attributes.count > 0 {
             // insert attributes
             for (key, value) in attributes {
-                xml += " \(key)=\"\(value)\""
+                xml += " \(key)=\"\(value.xmlEscaped)\""
             }
         }
         
@@ -262,16 +260,40 @@ public class AEXMLElement: NSObject {
                 xml += "</\(name)>"
             } else {
                 // insert string value and close element
-                xml += ">\(escapedStringValue)</\(name)>"
+                xml += ">\(stringValue.xmlEscaped)</\(name)>"
             }
         }
         
         return xml
     }
-
+    
+    /// Same as `xmlString` but without `\n` and `\t` characters
+    public var xmlStringCompact: String {
+        let chars = NSCharacterSet(charactersInString: "\n\t")
+        return xmlString.componentsSeparatedByCharactersInSet(chars).joinWithSeparator("")
+    }
+    
 }
 
-// MARK: -
+public extension String {
+    
+    /// String representation of self with XML special characters escaped.
+    public var xmlEscaped: String {
+        // we need to make sure "&" is escaped first. Not doing this may break escaping the other characters
+        var escaped = stringByReplacingOccurrencesOfString("&", withString: "&amp;", options: .LiteralSearch)
+        
+        // replace the other four special characters
+        let escapeChars = ["<" : "&lt;", ">" : "&gt;", "'" : "&apos;", "\"" : "&quot;"]
+        for (char, echar) in escapeChars {
+            escaped = escaped.stringByReplacingOccurrencesOfString(char, withString: echar, options: .LiteralSearch)
+        }
+        
+        return escaped
+    }
+    
+}
+
+// MARK: - AEXMLDocument
 
 /**
     This class is inherited from `AEXMLElement` and has a few addons to represent **XML Document**.
@@ -289,9 +311,11 @@ public class AEXMLDocument: AEXMLElement {
     
     /// Default options used by NSXMLParser
     public struct NSXMLParserOptions {
-        var shouldProcessNamespaces = false
-        var shouldReportNamespacePrefixes = false
-        var shouldResolveExternalEntities = false
+        public var shouldProcessNamespaces = false
+        public var shouldReportNamespacePrefixes = false
+        public var shouldResolveExternalEntities = false
+        
+        public init() {}
     }
     
     // MARK: Properties
@@ -308,22 +332,28 @@ public class AEXMLDocument: AEXMLElement {
     /// Options for NSXMLParser (default values are `false`)
     public let xmlParserOptions: NSXMLParserOptions
     
-    /// Root (the first child element) element of XML Document **(AEXMLError element if not exists)**.
-    private let errorElement = AEXMLElement(AEXMLElement.errorElementName, value: "XML Document must have root element.")
-    public var root: AEXMLElement { return children.count == 1 ? children.first! : errorElement }
+    /// Root (the first child element) element of XML Document **(Empty element with error if not exists)**.
+    public var root: AEXMLElement {
+        guard let rootElement = children.first else {
+            let errorElement = AEXMLElement()
+            errorElement.error = Error.RootElementMissing
+            return errorElement
+        }
+        return rootElement
+    }
     
     // MARK: Lifecycle
     
     /**
         Designated initializer - Creates and returns XML Document object.
     
-        :param: version Version value for XML Document header (defaults to 1.0).
-        :param: encoding Encoding value for XML Document header (defaults to "utf-8").
-        :param: standalone Standalone value for XML Document header (defaults to "no").
-        :param: root Root XML element for XML Document (defaults to `nil`).
-        :param: xmlParserOptions Options for NSXMLParser (defaults to `false` for all).
+        - parameter version: Version value for XML Document header (defaults to 1.0).
+        - parameter encoding: Encoding value for XML Document header (defaults to "utf-8").
+        - parameter standalone: Standalone value for XML Document header (defaults to "no").
+        - parameter root: Root XML element for XML Document (defaults to `nil`).
+        - parameter xmlParserOptions: Options for NSXMLParser (defaults to `false` for all).
     
-        :returns: An initialized XML Document object.
+        - returns: An initialized XML Document object.
     */
     public init(version: Double = Defaults.version,
                 encoding: String = Defaults.encoding,
@@ -352,13 +382,13 @@ public class AEXMLDocument: AEXMLElement {
     /**
         Convenience initializer - used for parsing XML data (by calling `loadXMLData:` internally).
     
-        :param: version Version value for XML Document header (defaults to 1.0).
-        :param: encoding Encoding value for XML Document header (defaults to "utf-8").
-        :param: standalone Standalone value for XML Document header (defaults to "no").
-        :param: xmlData XML data to parse.
-        :param: xmlParserOptions Options for NSXMLParser (defaults to `false` for all).
+        - parameter version: Version value for XML Document header (defaults to 1.0).
+        - parameter encoding: Encoding value for XML Document header (defaults to "utf-8").
+        - parameter standalone: Standalone value for XML Document header (defaults to "no").
+        - parameter xmlData: XML data to parse.
+        - parameter xmlParserOptions: Options for NSXMLParser (defaults to `false` for all).
     
-        :returns: An initialized XML Document object containing the parsed data. Throws error if the data could not be parsed.
+        - returns: An initialized XML Document object containing parsed data. Throws error if data could not be parsed.
     */
     public convenience init(version: Double = Defaults.version,
                             encoding: String = Defaults.encoding,
@@ -373,11 +403,10 @@ public class AEXMLDocument: AEXMLElement {
     // MARK: Read XML
     
     /**
-        Creates instance of `AEXMLParser` (private class which is simple wrapper around `NSXMLParser`) and starts parsing the given XML data.
+        Creates instance of `AEXMLParser` (private class which is simple wrapper around `NSXMLParser`) 
+        and starts parsing the given XML data. Throws error if data could not be parsed.
     
-        :param: data XML which should be parsed.
-    
-        :returns: `NSError` if parsing is not successfull, otherwise `nil`.
+        - parameter data: XML which should be parsed.
     */
     public func loadXMLData(data: NSData) throws {
         children.removeAll(keepCapacity: false)
@@ -398,7 +427,7 @@ public class AEXMLDocument: AEXMLElement {
     
 }
 
-// MARK: -
+// MARK: - AEXMLParser
 
 private class AEXMLParser: NSObject, NSXMLParserDelegate {